encrypting passwords and minor updates

+++ src/app.ts
@@ -74,24 +74,59 @@ const deletePassword = async (req: Request<delparam,delparam,delparam>, res: Res
 app.get("/",isAllowed, async (req: Request, res: Response) => {
   try{
    interface UserData {
-    createdAt: number;
      id: string;
      appname: string;
      password: string;
    }
 const userId = (req.session as any).user._id;
   const userData : string | null = await db.get(`userId:${userId}`);
+  //check if user was cached
 if (userData) {
   // User exists in Redis
   const parsedData: UserData[] = JSON.parse(userData);
- return res.render("index",{data:parsedData,recent:parsedData[0]});
+  //decrypt from redis
+   parsedData.map((current:any,index,array)=>{
+    const [encrypted,IV,authTag] = current.password.split(":")
+    const key = Buffer.from(process.env.ENCRYPTION_KEY!, "base64");
+    const decipher = crypto.createDecipheriv("aes-256-gcm"
+      ,key
+      ,Buffer.from(IV, "base64")
+    )
+    decipher.setAuthTag(Buffer.from(authTag, "base64"));
+    const decrypted = Buffer.concat([
+      decipher.update(Buffer.from(encrypted, "base64")),
+      decipher.final()
+    ])
+    current.password = decrypted.toString("utf8")
+    return current 
+    })
+    console.log((req.session as any).user)
+ return res.render("index",{data:parsedData,recent:parsedData[0],username:(req.session as any).user.username});
 }
 const passwords = await passwordModel.find({ id: userId }).sort({_id:-1})
+//if no passwrods
 if(passwords.length===0){
   return res.render("index",{data:[],recent:{appname:"AppName",password:"Password"}})  
 }
 await db.set(`userId:${userId}`, JSON.stringify(passwords),{EX:600});//after 10 minutes
-return res.render("index",{data:passwords,recent:passwords[0]})
+//decrypting
+ passwords.map((current:any,index,array)=>{
+const [encrypted,IV,authTag] = current.password.split(":")
+const key = Buffer.from(process.env.ENCRYPTION_KEY!, "base64");
+const decipher = crypto.createDecipheriv("aes-256-gcm"
+  ,key
+  ,Buffer.from(IV, "base64")
+)
+decipher.setAuthTag(Buffer.from(authTag, "base64"));
+const decrypted = Buffer.concat([
+  decipher.update(Buffer.from(encrypted, "base64")),
+  decipher.final()
+])
+current.password = decrypted.toString("utf8")
+return current 
+})
+//rendering
+return res.render("index",{data:passwords,recent:passwords[0],username:(req.session as any).user.username})
   }catch(err){
   console.error("Error fetching user data:", err);
   return res.status(500).send("Internal Server Error");
@@ -115,12 +150,25 @@ const parsedLength = parseInt(length);
  }
  //generate password
  const password : string = crypto.randomBytes(Math.ceil(parsedLength / 2)).toString("hex").slice(0, parsedLength);
+ //encrypting
+ //key
+ const key = Buffer.from(process.env.ENCRYPTION_KEY!, "base64");
+ const IV = crypto.randomBytes(12);
+ const cypher = crypto.createCipheriv("aes-256-gcm",key,IV) ;
+ //encrypting 
+ const encrypted = Buffer.concat([
+  cypher.update(password, "utf8"),
+  cypher.final()
+]);
+const getAuthTag = cypher.getAuthTag()
+const encryptedData = `${encrypted.toString("base64")}:${IV.toString("base64")}:${cypher.getAuthTag().toString("base64")}`;
+
  //storing password
 await passwordModel.insertOne(
   {
     id: userId,
     appname: appName,
-    password
+    password:encryptedData
   }
 );
  //deleting caching
@@ -134,6 +182,9 @@ await passwordModel.insertOne(
   try{
   //getting user id
   const userId = (req.session as any).user._id;
+  //checking if there is passwords
+  const passwords = await passwordModel.find({id:userId})
+  if(passwords.length===0) return res.status(400).json({msg:"no data to clear"})
   //deleting all passwords associated with that id
   await passwordModel.deleteMany({id:userId})
   //removing cache
